// src/domains/dam/types/Asset.ts
import type { DocId, DocIdNullable, IntegerId } from '@/shared/types/common'
import type { AnzuUserAndTimeTrackingAware } from '@/shared/types/AnzuUserAndTimeTrackingAware'
import type { ResourceNameSystemAware } from '@/shared/types/ResourceNameSystemAware'
import type { AssetFile } from '@/domains/dam/types/AssetFile'

interface Texts {
  displayTitle: string
}

export type DamDistributionServiceName = string

export const DamAssetStatus = {
  Draft: 'draft',
  WithFile: 'with_file',
  Deleting: 'deleting',
} as const
export type DamAssetStatusType = (typeof DamAssetStatus)[keyof typeof DamAssetStatus]
export const DamAssetStatusDefault = DamAssetStatus.Draft

export const DamAssetType = {
  Image: 'image',
  Audio: 'audio',
  Video: 'video',
  Document: 'document',
} as const
export type DamAssetTypeType = (typeof DamAssetType)[keyof typeof DamAssetType]
export const DamAssetTypeDefault = DamAssetType.Image

interface Attributes {
  assetType: DamAssetTypeType
  assetStatus: DamAssetStatusType
}

export interface AssetFileProperties {
  distributesInServices: DamDistributionServiceName[]
  slotNames: string[]
  fromRss: boolean
  width: number
  height: number
}

interface Flags {
  described: boolean
  visible: boolean
}

export interface AssetMetadataSuggestions extends Record<string, Array<string>> {}

export type AssetCustomData = Record<string, any>

interface Metadata {
  authorSuggestions: AssetMetadataSuggestions
  keywordSuggestions: AssetMetadataSuggestions
  customData: any
}

export interface AssetSearchListItemDto extends AnzuUserAndTimeTrackingAware, ResourceNameSystemAware {
  id: DocId
  texts: Texts
  attributes: Attributes
  flags: Flags
  licence: IntegerId
  mainFile: null | AssetFile
  podcasts: DocId[]
  assetFileProperties: AssetFileProperties
  mainFileSingleUse: Readonly<boolean | null>
  mainFileInternal: Readonly<boolean | null>
  mainFileOverrideInternal: Readonly<boolean | null>
}

export interface AssetDetailItemDto extends AnzuUserAndTimeTrackingAware, ResourceNameSystemAware {
  id: DocId
  texts: Texts
  attributes: Attributes
  flags: Flags
  licence: IntegerId
  mainFile: null | AssetFile
  keywords: DocId[]
  authors: DocId[]
  podcasts: DocId[]
  metadata: Metadata
  siblingToAsset: DocIdNullable
  distributionCategory: DocIdNullable
  assetFileProperties: AssetFileProperties
  mainFileSingleUse: Readonly<boolean | null>
  mainFileInternal: Readonly<boolean | null>
  mainFileOverrideInternal: Readonly<boolean | null>
}

export interface AssetMetadataDto extends AnzuUserAndTimeTrackingAware, ResourceNameSystemAware, Metadata {
  id: DocId
  customData: AssetCustomData
}

export type AssetExternalProviderId = string | number
export type AssetExternalProviderIdNullable = AssetExternalProviderId | null

export type AssetExternalProviderMetadata = Record<string, string | number | number[] | string[] | boolean>

export interface DamImageCopyToLicenceRequestItem {
  asset: DocId
  targetAssetLicence: IntegerId
}

export type DamImageCopyToLicenceRequest = Array<DamImageCopyToLicenceRequestItem>

export interface DamImageCopyToLicenceResponseItem {
  asset: DocId
  targetAsset: DocId
  targetMainFile: DocId
  targetAssetLicence: IntegerId
  result: 'exists' | 'copy' | 'notAllowed' | 'unassigned'
  assetConflicts: DocId[]
}

export type DamImageCopyToLicenceResponse = Array<DamImageCopyToLicenceResponseItem>